% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aston_villa}
\alias{aston_villa}
\title{Futbin database at 12th March 2020.}
\format{A data frame with 21705 rows and 22 variables:
\describe{
  \item{name}{price, in US dollars}
  \item{rating}{weight of the diamond, in carats}
  \item{position}{weight of the diamond, in carats}
  \item{ver}{weight of the diamond, in carats}
  \item{ps_price}{weight of the diamond, in carats}
  \item{skills}{weight of the diamond, in carats}
  \item{weak_foot}{weight of the diamond, in carats}
  \item{work_rate}{weight of the diamond, in carats}
  \item{carat}{weight of the diamond, in carats}
  \item{carat}{weight of the diamond, in carats}
  \item{carat}{weight of the diamond, in carats}
  \item{carat}{weight of the diamond, in carats}
  \item{carat}{weight of the diamond, in carats}
  \item{carat}{weight of the diamond, in carats}
  \item{carat}{weight of the diamond, in carats}
  \item{carat}{weight of the diamond, in carats}
  \item{carat}{weight of the diamond, in carats}
  \item{carat}{weight of the diamond, in carats}
}}
\source{
\url{http://www.futbin.com/}
}
\usage{
aston_villa
}
\description{
A dataset containing the stats available on Futbin of Aston Villa players.
}
\details{
Generated with:
futbin_scrap(url = "https://www.futbin.com/players?page=1&club=2")
}
\keyword{datasets}
