% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futbin_plot.R
\name{futbin_plot}
\alias{futbin_plot}
\title{futbin_plot makes an interactive radar plot of the stats of the players.}
\usage{
futbin_plot(df, gk = FALSE)
}
\arguments{
\item{df}{dataframe generated with columns `pace`, `sho`, `pas`, `dri`, `def`, `phy`,
`name`, `rating` and `version`.
This dataframe can be obtained from function \code{\link{futbin_search}}.}

\item{gk}{Optional. If `TRUE`, the labels of the plot are the stats
for goalkeepers: diving, handling, kicking, reflexes, speed and position.}
}
\value{
An interactive radar plot of the stats.
}
\description{
futbin_plot makes an interactive radar plot of the stats of the players.
}
\examples{

\dontrun{
# Plot comparing Van Dijk and Messi
players <- futbin_search(name = c("Van Dijk", "Lionel Messi"), version = "Rare")
futbin_plot(players)

# Plot comparing goalkeepers
some_goalkeepers <- futbin_search(name = c("De Gea", "Kepa", "Hugo Lloris"), version = "Rare")
futbin_plot(some_goalkeepers, gk = TRUE)
}
}
