% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futbin_scrap.R
\name{futbin_scrap}
\alias{futbin_scrap}
\title{futbin_scrap extract players in a URL in Futbin.}
\usage{
futbin_scrap(url, sleep_time = 5, messages = TRUE)
}
\arguments{
\item{url}{Futbin URL to web scrap. All pages of the URL will be automatically detected and downloaded.
Futbin webpage (futbin.com/players) can be used to make customised filters, and then copy the URL here.}

\item{sleep_time}{Time (in seconds) ellapsed between one download and the next. Please respect Futbin API.}

\item{messages}{Optional. To show additional messages about webpage used and number of players found.}
}
\value{
A dataframe with all the players found searching for \code{name} and  \code{version}.
}
\description{
futbin_scrap extract players in a URL in Futbin.
}
\examples{
\dontrun{
# Aston Villa players
futbin_scrap(url = "https://www.futbin.com/players?page=1&club=2")

# Spanish centre-backs from La Liga
futbin_scrap(url = "https://www.futbin.com/players?page=1&nation=45&league=53&position=CB")
}
}
