% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futbin_search.R
\name{futbin_search}
\alias{futbin_search}
\title{futbin_search searchs players in futbin.}
\usage{
futbin_search(name = "", version_selected = NULL, verbose = F)
}
\arguments{
\item{name}{Optional. Vector with the names of the players. If not specified
it will report the 30 highest-rated players of the game.}

\item{version_selected}{Optional. Version of the cards. Some options are "Normal",
"CL" (Champions League), "IF" (In-Form), "SIF" (Second In-Form), ...}

\item{verbose}{Optional. To show additional verbose about webpage used and number of players found.}
}
\value{
A dataframe with all the players found searching for \code{name} and  \code{version}.
}
\description{
futbin_search searchs players in futbin.
}
\examples{
# Search for a player
futbin_search(name = "Lionel Messi")
# Search for more than one player
futbin_search(name = c("Lionel Messi", "Cristiano Ronaldo"))
# Search for a specific version of two or more players
futbin_search(name = c("Lionel Messi", "Griezmann"), version_selected = "Normal")
# Search for an In-Form card of a player, showing verbose
futbin_search(name = "Grealish", version_selected = "IF", verbose = TRUE)
}
