% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futbin_search.R
\name{futbin_search}
\alias{futbin_search}
\title{futbin_search searchs players in futbin.}
\usage{
futbin_search(name = "", version = NULL, messages = F)
}
\arguments{
\item{name}{Optional. Vector with the names of the players. If not specified
it will report the 30 highest-rated players of the game.}

\item{version}{Optional. Version of the cards. Some options are "Normal",
"CL" (Champions League), "IF" (In-Form), "SIF" (Second In-Form), ...}

\item{messages}{Optional. To show additional messages about webpage used and number of players found.}
}
\value{
A dataframe with all the players found searching for \code{name} and  \code{version}.
}
\description{
futbin_search searchs players in futbin.
}
\examples{
futbin_search(name = "Lionel Messi")
futbin_search(name = c("Lionel Messi", "Griezmann"), version = "Normal")
futbin_search(name = "Luka Modric", version = "IF", messages = TRUE)
}
