% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{futbin_search}
\alias{futbin_search}
\title{futbin_search searchs players in futbin.}
\usage{
futbin_search(name = "", version = NULL, messages = F)
}
\arguments{
\item{name}{Optional. Name of the player. If leaved blank,
it will report the 30 highest-rated players of the game.}

\item{version}{Optional. Version of the card. It can be "Normal",
"CL" (Champions League), "IF" (In-Form), "SIF" (Second In-Form), ...}

\item{messages}{Optional. To show additional messages
(webpage used and number of players found)}
}
\value{
A dataframe with all the players found searching for \code{name} and  \code{version}
}
\description{
futbin_search searchs players in futbin.
}
\examples{
futbin_search("Lionel Messi")
futbin_search("Lionel Messi", version = "Normal")
futbin_search("Luka Modric", version = "IF", messages = TRUE)
}
